requirements.txt
google-cloud-storage==2.16.0
google-cloud-pubsub==2.21.1
flask==3.0.3
gunicorn==22.0.0
pytest==8.1.1
pytest-mock==3.14.0
